StreamlitApps/Jira/FirstResponseTime/requirements.txt
pandas
streamlit
python-decouple
openpyxl
mysqlclient